pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "gagnon florian",email = "gagnon.florian@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django (>=6.0,<7.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "ruff (>=0.14.7,<0.15.0)"
]

[tool.poetry]
packages = [{include = "backend"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"